+++ demos/session_service/src/main/java/androidx/media3/demo/session/DemoMediaLibrarySessionCallback.kt
@@ -21,6 +21,7 @@ import androidx.core.net.toUri
 import androidx.media3.common.C
 import androidx.media3.common.MediaItem
 import androidx.media3.common.MediaMetadata
+import androidx.media3.common.Player
 import androidx.media3.common.util.UnstableApi
 import androidx.media3.demo.session.service.R
 import androidx.media3.session.CommandButton
@@ -30,6 +31,7 @@ import androidx.media3.session.MediaLibraryService
 import androidx.media3.session.MediaSession
 import androidx.media3.session.MediaSession.MediaItemsWithStartPosition
 import androidx.media3.session.SessionCommand
+import androidx.media3.session.SessionCommands
 import androidx.media3.session.SessionError
 import androidx.media3.session.SessionResult
 import com.google.common.collect.ImmutableList
@@ -62,7 +64,7 @@ open class DemoMediaLibrarySessionCallback(val service: DemoPlaybackService) :
     )
 
   @OptIn(UnstableApi::class) // MediaSession.ConnectionResult.DEFAULT_SESSION_AND_LIBRARY_COMMANDS
-  val mediaNotificationSessionCommands =
+  private val sessionCommandsWithCustomCommands =
     MediaSession.ConnectionResult.DEFAULT_SESSION_AND_LIBRARY_COMMANDS.buildUpon()
       .also { builder ->
         // Put all custom session commands in the list that may be used by the notification.
@@ -72,27 +74,35 @@ open class DemoMediaLibrarySessionCallback(val service: DemoPlaybackService) :
       }
       .build()
 
-  // ConnectionResult.DEFAULT_SESSION_AND_LIBRARY_COMMANDS
-  // ConnectionResult.AcceptedResultBuilder
-  @OptIn(UnstableApi::class)
+  @OptIn(UnstableApi::class) // Player.Commands.Builder
+  private val restrictedAccessPlayerCommands =
+    Player.Commands.Builder()
+      .addAll(
+        Player.COMMAND_GET_CURRENT_MEDIA_ITEM,
+        Player.COMMAND_GET_TRACKS,
+        Player.COMMAND_GET_METADATA,
+      )
+      .build()
+
+  @OptIn(UnstableApi::class) // ConnectionResult.AcceptedResultBuilder
   override fun onConnect(
     session: MediaSession,
     controller: MediaSession.ControllerInfo,
   ): MediaSession.ConnectionResult {
-    if (
-      session.isMediaNotificationController(controller) ||
-        session.isAutomotiveController(controller) ||
-        session.isAutoCompanionController(controller)
-    ) {
-      // Select the button to display.
+    if (controller.isTrusted) {
+      // Provide full information and specify media button preferences.
       val customButton = commandButtons[if (session.player.shuffleModeEnabled) 1 else 0]
       return MediaSession.ConnectionResult.AcceptedResultBuilder(session)
-        .setAvailableSessionCommands(mediaNotificationSessionCommands)
+        .setAvailableSessionCommands(sessionCommandsWithCustomCommands)
         .setMediaButtonPreferences(ImmutableList.of(customButton))
         .build()
+    } else {
+      // Restricted read-only view on currently playing item only.
+      return MediaSession.ConnectionResult.accept(
+        SessionCommands.EMPTY,
+        restrictedAccessPlayerCommands,
+      )
     }
-    // Default commands without media button preferences for common controllers.
-    return MediaSession.ConnectionResult.AcceptedResultBuilder(session).build()
   }
 
   @OptIn(UnstableApi::class) // MediaSession.isMediaNotificationController