+++ libraries/session/src/main/java/androidx/media3/session/MediaSession.java
@@ -622,18 +622,19 @@ public class MediaSession {
     }
 
     /**
-     * Returns whether one or more of the following is true:
+     * Returns whether the controller is trusted by the user to control and access media.
+     *
+     * <p>One or more of the following must be true for the controller to be trusted:
      *
      * <ul>
-     *   <li>the controller's app UID is in the same UID as the current process
-     *   <li>the controller's app UID is the system UID
-     *   <li>the controller has been granted {@code android.permission.MEDIA_CONTENT_CONTROL}
-     *   <li>the controller has been granted {@code android.permission.STATUS_BAR_SERVICE}
-     *   <li>the controller has an enabled notification listener
+     *   <li>The controller is part of the current app and user (using {@link
+     *       android.os.Process#myUid()}.
+     *   <li>The controller is part of the Android system (using {@link
+     *       android.os.Process#SYSTEM_UID}.
+     *   <li>The controller has been granted {@code android.permission.MEDIA_CONTENT_CONTROL}.
+     *   <li>The controller has been granted {@code android.permission.STATUS_BAR_SERVICE}.
+     *   <li>The controller has an enabled notification listener.
      * </ul>
-     *
-     * to determine if the controller can be trusted to accept connection and incoming command
-     * requests.
      */
     @UnstableApi
     public boolean isTrusted() {