+++ RELEASENOTES.md
@@ -52,6 +52,9 @@
     *   Add new parameter to `MediaSession.Callback.onPlaybackResumption` to
         indicate if the call happens to gather information only or to start
         playback ([#1764](https://github.com/androidx/media/issues/1764)).
+    *   Update `MediaSession.ControllerInfo.isTrusted` to also declare
+        controllers from the own app as trusted
+        ([#2542](https://github.com/androidx/media/issues/2542)).
 *   UI:
     *   Fix bug where `PlayerSurface` inside re-usable components like
         `LazyColumn` didn't work correctly